Update docker-entrypoint.sh
The release nupkg build script was looking for release binaries, but we're trying to build debug binaries here, so we have to hack around it a bit. I left windows builds as release cause we don't need them to be debug currently, but this should build a Linux debug binary now.